requirements.txt
# Python dependencies for data processing and ingestion
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pymysql>=1.0.0
mysql-connector-python>=8.0.0
pyarrow>=11.0.0
fastparquet>=0.8.0
xlsxwriter>=3.0.0
python-dateutil>=2.8.0
cryptography>=3.4.0
